abstract out renderTables
